dom/workers/test/test_navigator_languages.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Navigator
-->
<head>
  <title>Test for DOM Worker Navigator.languages</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" language="javascript">

  var tests = [
    { expectedLanguages: 'en,it', inputLanguages: 'en,it' },
    { expectedLanguages: 'it,en,fr', inputLanguages: 'it,en,fr' },
    { expectedLanguages: SpecialPowers.Services.locale.webExposedLocales[0], inputLanguages: '' },
    { expectedLanguages: 'en,it', inputLanguages: 'en,it' },
  ];
  var test;
  function runTests() {
    if (!tests.length) {
      worker.postMessage('finish');
      SimpleTest.finish();
      return;
    }

    test = tests.shift();
    SpecialPowers.pushPrefEnv({"set": [["intl.accept_languages", test.inputLanguages]]}, function() {
      worker.postMessage(true);
    });
  }

  SimpleTest.waitForExplicitFinish();

  var worker = new Worker("navigator_languages_worker.js");

  worker.onmessage = function(event) {
    is(event.data.toString(), navigator.languages.toString(), "The languages match.");
    is(event.data.toString(), test.expectedLanguages, "This is the correct result.");
    runTests();
  }

  runTests();

</script>
</pre>
</body>
</html>
